requirements.txt
asgiref==3.8.1
asttokens==2.4.1
coverage==7.4.4
decorator==5.1.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
exceptiongroup==1.2.1
executing==2.0.1
ipython==8.23.0
jedi==0.19.1
mysqlclient==2.2.4
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
typing_extensions==4.11.0
wcwidth==0.2.13
